test: upgrade phpunit to 9.5
