+++ Tests/Functional/app/AppKernel.php
@@ -1,44 +1,29 @@
 <?php
 
+use Payum\Bundle\PayumBundle\Tests\Functional\app\AppKernelShared;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Kernel;
-use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 
-class AppKernel extends Kernel
+# Could be simplified after support of symfony 4 support is dropped
+if(Kernel::MAJOR_VERSION===4)
 {
-    public function registerBundles(): array
+    class AppKernel extends AppKernelShared
     {
-        $bundles = array(
-            new Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
-            new Symfony\Bundle\TwigBundle\TwigBundle(),
-            new Payum\Bundle\PayumBundle\PayumBundle(),
-        );
-
-        return $bundles;
+        public function handle(Request $request, $type = HttpKernelInterface::MAIN_REQUEST, $catch = true): Response
+        {
+            return parent::handle($request, $type, false);
+        }
     }
-
-    public function handle(Request $request, int $type = HttpKernelInterface::MASTER_REQUEST, bool $catch = true): Response
+}
+else
+{
+    class AppKernel extends AppKernelShared
     {
-        /** @noinspection SuspiciousAssignmentsInspection */
-        $catch = false;
-
-        return parent::handle($request, $type, $catch);
-    }
-
-    public function getCacheDir(): string
-    {
-        return sys_get_temp_dir() . '/PayumBundle/cache';
-    }
-
-    public function getLogDir(): string
-    {
-        return sys_get_temp_dir() . '/PayumBundle/logs';
-    }
-
-    public function registerContainerConfiguration(LoaderInterface $loader): void
-    {
-        $loader->load(__DIR__ . '/config/config.yml');
+        public function handle(Request $request, int $type = HttpKernelInterface::MAIN_REQUEST, bool $catch = true): Response
+        {
+            return parent::handle($request, $type, false);
+        }
     }
 }
